store creation and file restructure
