--- src/components/MovieIndex.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-export default class MovieIndex extends Component {
-  constructor() {
-    super();
-  }
-
-  render() {
-    return (
-      <div>
-      MOVIES
-      </div>
-    )
-  }
-}
