+++ src/components/index.js
@@ -1,0 +1,5 @@
+import SignIn from './SignIn/SignIn';
+import MovieCard from './MovieCard/MovieCard';
+import MovieContainer from './MovieContainer/MovieContainer';
+
+export { SignIn , MovieContainer, MovieCard };
