+++ src/index.js
@@ -1,17 +1,15 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App/App';
-import MovieIndex from './components/MovieIndex';
+//import MovieIndex from './components/MovieIndex';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import rootReducer from './reducers';
-import configureStore from './store/configureStore';
-
-
-const store = configureStore();
+import rootReducer from './Reducers';
+//import configureStore from './store/configureStore';
+import Store from './Store';
 
 ReactDOM.render(
-  <Provider store={ store }>
+  <Provider store={ Store }>
     <App/>
   </Provider>,
   document.getElementById('root')
