+++ src/reducers/moviesIsLoading.js
@@ -1,6 +1,6 @@
 const moviesIsLoading = (state = false, action) => {
   switch (action.type) {
-    case 'MOVIESS_IS_LOADING':
+    case 'MOVIES_IS_LOADING':
       return action.isLoading;
 
     default:
